Support HexText raw mode highlighting
